requirements.txt
streamlit
streamlit-cropper
torch
torchvision
pillow
ultralytics
sentence-transformers
transformers
timm
numpy